+++ src/Suru.Compiler/Parse/Ast/ReturnStatement.cs
@@ -1,0 +1,6 @@
+namespace Suru.Compiler.Parse.Ast;
+
+public sealed class ReturnStatement(Expression? value) : Statement
+{
+    public Expression? Value { get; } = value;
+}